Update File I/O in fs2.io.file.
 - Move FileHandle creation methods to FileHandle object.
 - Methods that return a `Pull` move to `fs2.io.file.pulls`.
 - Can create a `Pull` from `Path`, `FileChannel`, or `AsynchronousFileChannel`.
 - Add some Markdown documentation for the code.
