+++ io/src/main/scala/fs2/io/file/FileHandle.scala
@@ -1,6 +1,10 @@
 package fs2.io.file
 
-import fs2.Chunk
+import java.nio.ByteBuffer
+import java.nio.channels.{AsynchronousFileChannel, FileChannel, FileLock}
+
+import fs2._
+import fs2.util.Monad
 
 trait FileHandle[F[_]] {
   type Lock
@@ -79,4 +83,108 @@ trait FileHandle[F[_]] {
     * @return the number of bytes written.
     */
   def write(bytes: Chunk[Byte], offset: Long): F[Int]
+}
+
+object FileHandle {
+  /**
+    * Creates a `FileHandle[F]` from a `java.nio.channels.AsynchronousFileChannel`.
+    *
+    * Uses a `java.nio.Channels.CompletionHandler` to handle callbacks from IO operations.
+    */
+  private[fs2] def fromAsynchronousFileChannel[F[_]](chan: AsynchronousFileChannel)(implicit F: Async[F], S: Strategy): FileHandle[F] = {
+    new FileHandle[F] {
+      type Lock = FileLock
+
+      override def close(): F[Unit] =
+        F.suspend(chan.close())
+
+      override def force(metaData: Boolean): F[Unit] =
+        F.suspend(chan.force(metaData))
+
+      override def lock: F[Lock] =
+        asyncCompletionHandler[F, Lock](f => F.pure(chan.lock(null, f)))
+
+      override def lock(position: Long, size: Long, shared: Boolean): F[Lock] =
+        asyncCompletionHandler[F, Lock](f => F.pure(chan.lock(position, size, shared, null, f)))
+
+      override def read(numBytes: Int, offset: Long): F[Option[Chunk[Byte]]] = {
+        val buf = ByteBuffer.allocate(numBytes)
+        F.bind(
+          asyncCompletionHandler[F, Integer](f => F.pure(chan.read(buf, offset, null, f))))(
+          len => F.pure {
+            if (len < 0) None else if (len == 0) Some(Chunk.empty) else Some(Chunk.bytes(buf.array.take(len)))
+          }
+        )
+      }
+
+      override def size: F[Long] =
+        F.suspend(chan.size)
+
+      override def truncate(size: Long): F[Unit] =
+        F.suspend { chan.truncate(size); () }
+
+      override def tryLock: F[Option[Lock]] =
+        F.map(F.suspend(chan.tryLock()))(Option.apply)
+
+      override def tryLock(position: Long, size: Long, shared: Boolean): F[Option[Lock]] =
+        F.map(F.suspend(chan.tryLock(position, size, shared)))(Option.apply)
+
+      override def unlock(f: Lock): F[Unit] =
+        F.suspend(f.release())
+
+      override def write(bytes: Chunk[Byte], offset: Long): F[Int] =
+        F.map(
+          asyncCompletionHandler[F, Integer](f => F.pure(chan.write(ByteBuffer.wrap(bytes.toArray), offset, null, f))))(
+          i => i.toInt
+        )
+    }
+  }
+
+  /**
+    * Creates a `FileHandle[F]` from a `java.nio.channels.FileChannel`.
+    */
+  private[fs2] def fromFileChannel[F[_]](chan: FileChannel)(implicit F: Monad[F]): FileHandle[F] = {
+    new FileHandle[F] {
+      type Lock = FileLock
+
+      override def close(): F[Unit] =
+        F.suspend(chan.close())
+
+      override def force(metaData: Boolean): F[Unit] =
+        F.suspend(chan.force(metaData))
+
+      override def lock: F[Lock] =
+        F.suspend(chan.lock)
+
+      override def lock(position: Long, size: Long, shared: Boolean): F[Lock] =
+        F.suspend(chan.lock(position, size, shared))
+
+      override def read(numBytes: Int, offset: Long): F[Option[Chunk[Byte]]] = {
+        val buf = ByteBuffer.allocate(numBytes)
+        F.bind(
+          F.suspend(chan.read(buf, offset)))(len => F.pure {
+          if (len < 0) None else if (len == 0) Some(Chunk.empty) else Some(Chunk.bytes(buf.array.take(len)))
+        }
+        )
+      }
+
+      override def size: F[Long] =
+        F.suspend(chan.size)
+
+      override def truncate(size: Long): F[Unit] =
+        F.suspend { chan.truncate(size); () }
+
+      override def tryLock: F[Option[Lock]] =
+        F.suspend(Option(chan.tryLock()))
+
+      override def tryLock(position: Long, size: Long, shared: Boolean): F[Option[Lock]] =
+        F.suspend(Option(chan.tryLock(position, size, shared)))
+
+      override def unlock(f: Lock): F[Unit] =
+        F.suspend(f.release())
+
+      override def write(bytes: Chunk[Byte], offset: Long): F[Int] =
+        F.suspend(chan.write(ByteBuffer.wrap(bytes.toArray), offset))
+    }
+  }
 }
